requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.28.2
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
persiantools>=3.0.1
apscheduler>=3.10.1
pydantic>=1.10.7
python-multipart>=0.0.6
jinja2>=3.1.2
loguru>=0.7.0 